+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project are documented here.
 This project adheres to [Semantic Versioning](https://semver.org/).
 
-## [Unreleased]
+## [1.5.2] — 2026-08-09
 
 ### Changed